requirements.txt
# --- Core ---
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

numpy>=1.23.0
opencv-python>=4.7.0
Pillow>=9.4.0
tqdm>=4.64.0

# --- Image Processing / Plotting ---
scikit-image>=0.20.0
matplotlib>=3.6.0

# --- YAML config parsing ---
pyyaml>=6.0

# --- Metrics ---
lpips>=0.1.4

# --- FLOPs / Params measurement ---
thop>=0.1.1.post2209072238
calflops>=0.2.0

# --- Logging ---
tensorboard>=2.10.0

# --- Misc Utilities ---
scipy>=1.10.0